+++ src/components/Filters/Filters.types.ts
@@ -5,6 +5,8 @@ import { Operators } from './Filters.enums';
 import {
   DateRange,
   DateRangePickerPropTypes,
+  DateRangePlaceholderPropTypes,
+  DateRangePlaceholderProps,
 } from './components/DateRangePicker/DateRangePicker.types';
 
 type OperatorTypes = typeof Operators[keyof typeof Operators];
@@ -20,6 +22,8 @@ interface ComponentProps {
   patternMessage?: string;
   minDate?: Date;
   maxDate?: Date;
+  units?: string;
+  placeholder?: string | DateRangePlaceholderProps;
 }
 
 export interface ComponentWithProps {
@@ -77,6 +81,11 @@ export const FieldPropTypes = PropTypes.exact({
             patternMessage: PropTypes.string,
             minDate: PropTypes.instanceOf(Date),
             maxDate: PropTypes.instanceOf(Date),
+            units: PropTypes.string,
+            placeholder: PropTypes.oneOfType([
+              PropTypes.string,
+              DateRangePlaceholderPropTypes,
+            ]),
           }),
         }),
       ]).isRequired,