+++ src/components/Filters/components/DateRangePicker/DateRangePicker.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import styled from 'styled-components';
 import DatePicker from 'react-datepicker';
-import { any } from 'ramda';
+import { any, prop } from 'ramda';
 import { isNotNull } from 'ramda-adjunct';
 
 import { FlexContainer } from '../../../FlexContainer';
@@ -10,6 +10,7 @@ import { datePickerStyles, dateRangePickerStyles } from './styles';
 import {
   DateRangePickerPropTypes,
   DateRangePickerProps,
+  DateRangePlaceholderPropTypes,
 } from './DateRangePicker.types';
 
 const StyledDatePicker = styled(FlexContainer)`
@@ -24,8 +25,11 @@ const DateRangePicker: React.FC<DateRangePickerProps> = ({
   onChange,
   minDate,
   maxDate,
+  placeholder,
 }) => {
   const { startDate, endDate } = value;
+  const startDatePlaceholder = prop('startDate', placeholder);
+  const endDatePlaceholder = prop('endDate', placeholder);
 
   const handleStartDateChange = (newStartDate) => {
     onChange(
@@ -52,7 +56,7 @@ const DateRangePicker: React.FC<DateRangePickerProps> = ({
         endDate={endDate}
         maxDate={maxDate}
         minDate={minDate}
-        placeholderText="Start date"
+        placeholderText={startDatePlaceholder || 'Start date'}
         selected={startDate}
         startDate={startDate}
         selectsStart
@@ -65,7 +69,7 @@ const DateRangePicker: React.FC<DateRangePickerProps> = ({
         endDate={endDate}
         maxDate={maxDate}
         minDate={startDate || minDate}
-        placeholderText="End date"
+        placeholderText={endDatePlaceholder || 'End date'}
         selected={endDate}
         startDate={startDate}
         selectsEnd
@@ -82,4 +86,5 @@ DateRangePicker.propTypes = {
   value: DateRangePickerPropTypes,
   minDate: PropTypes.instanceOf(Date),
   maxDate: PropTypes.instanceOf(Date),
+  placeholder: DateRangePlaceholderPropTypes,
 };