+++ src/components/Filters/components/Input/Input.types.ts
@@ -1,6 +1,7 @@
 export interface InputProps {
   value: string;
   onChange: (event: React.ReactEventHandler) => void;
+  placeholder?: string;
   isInvalid?: boolean;
   maxLength?: number;
   pattern?: string;