+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+# [1.0.0-alpha.51](https://github.com/securityscorecard/design-system/compare/v1.0.0-alpha.50...v1.0.0-alpha.51) (2021-04-09)
+
+
+### Features
+
+* **DateRangePicker:** add custom placeholder ([e0f95e5](https://github.com/securityscorecard/design-system/commit/e0f95e52d5ec7b0ca27d115a1cf72c0edee6cb54))
+* **Filters:** add custom placeholder and units ([b3ef288](https://github.com/securityscorecard/design-system/commit/b3ef288703cb2e2281afe43972729829a0aa9472))
+* **Storybook:** update Filters ([3b3620c](https://github.com/securityscorecard/design-system/commit/3b3620c965bc466a5cc40c6bae2c9e0642b4d695))
+
+# [1.0.0-alpha.50](https://github.com/securityscorecard/design-system/compare/v1.0.0-alpha.49...v1.0.0-alpha.50) (2021-04-08)
+
+
+### Features
+
+* **Datatable:** show scrollbar above table ([3327de2](https://github.com/securityscorecard/design-system/commit/3327de24d29c4862cf0799302aa383e1baf02340))
+
 # [1.0.0-alpha.49](https://github.com/securityscorecard/design-system/compare/v1.0.0-alpha.48...v1.0.0-alpha.49) (2021-04-06)